计算的时候int改为double

+++ ComeCapture/Controls/MainImage.cs
@@ -167,8 +167,8 @@ namespace ComeCapture.Controls
         #region 滑动中事件
         private void OnDragDelta(object sender, DragDeltaEventArgs e)
         {
-            var X = (int)e.HorizontalChange;
-            var Y = (int)e.VerticalChange;
+            var X = e.HorizontalChange;
+            var Y = e.VerticalChange;
             switch (Direction)
             {
                 case Direction.Move:
@@ -243,7 +243,7 @@ namespace ComeCapture.Controls
                     case Tool.Rectangle:
                         if (_Rectangle != null)
                         {
-                            ResetLimit((int)Canvas.GetLeft(_Rectangle), (int)Canvas.GetTop(_Rectangle), (int)Canvas.GetLeft(_Rectangle) + (int)_Rectangle.Width, (int)Canvas.GetTop(_Rectangle) + (int)_Rectangle.Height);
+                            ResetLimit(Canvas.GetLeft(_Rectangle), Canvas.GetTop(_Rectangle), Canvas.GetLeft(_Rectangle) + _Rectangle.Width, Canvas.GetTop(_Rectangle) + _Rectangle.Height);
                             MainWindow.Register(_Rectangle);
                             _Rectangle = null;
                         }
@@ -251,7 +251,7 @@ namespace ComeCapture.Controls
                     case Tool.Ellipse:
                         if (_Ellipse != null)
                         {
-                            ResetLimit((int)Canvas.GetLeft(_Ellipse), (int)Canvas.GetTop(_Ellipse), (int)Canvas.GetLeft(_Ellipse) + (int)_Ellipse.Width, (int)Canvas.GetTop(_Ellipse) + (int)_Ellipse.Height);
+                            ResetLimit(Canvas.GetLeft(_Ellipse), Canvas.GetTop(_Ellipse), Canvas.GetLeft(_Ellipse) + _Ellipse.Width, Canvas.GetTop(_Ellipse) + _Ellipse.Height);
                             MainWindow.Register(_Ellipse);
                             _Ellipse = null;
                         }
@@ -260,7 +260,7 @@ namespace ComeCapture.Controls
                         if (_Arrow != null)
                         {
                             geometry.Clear();
-                            ResetLimit((int)points.Min(p => p.X), (int)points.Min(p => p.Y), (int)points.Max(p => p.X), (int)points.Max(p => p.Y));
+                            ResetLimit(points.Min(p => p.X), points.Min(p => p.Y), points.Max(p => p.X), points.Max(p => p.Y));
                             points = null;
                             MainWindow.Register(_Arrow);
                             _Arrow = null;
@@ -270,7 +270,7 @@ namespace ComeCapture.Controls
                         if (_Line != null)
                         {
                             geometry.Clear();
-                            ResetLimit((int)points.Min(p => p.X), (int)points.Min(p => p.Y), (int)points.Max(p => p.X), (int)points.Max(p => p.Y));
+                            ResetLimit(points.Min(p => p.X), points.Min(p => p.Y), points.Max(p => p.X), points.Max(p => p.Y));
                             points = null;
                             MainWindow.Register(_Line);
                             _Line = null;
@@ -287,7 +287,7 @@ namespace ComeCapture.Controls
         #endregion
 
         #region 画矩形
-        private void DrawRectangle(int X, int Y)
+        private void DrawRectangle(double X, double Y)
         {
             if (_Rectangle == null)
             {
@@ -302,29 +302,29 @@ namespace ComeCapture.Controls
             }
             if (X > 0)
             {
-                Canvas.SetLeft(_Rectangle, (int)point.X + AppModel.Current.MaskLeftWidth);
-                _Rectangle.Width = X < Width - (int)point.X ? X : (int)Width - (int)point.X;
+                Canvas.SetLeft(_Rectangle, point.X + AppModel.Current.MaskLeftWidth);
+                _Rectangle.Width = X < Width - point.X ? X : Width - point.X;
             }
             else
             {
-                Canvas.SetLeft(_Rectangle, -X < point.X ? (int)point.X + X + AppModel.Current.MaskLeftWidth : AppModel.Current.MaskLeftWidth);
-                _Rectangle.Width = -X < point.X ? -X : (int)point.X;
+                Canvas.SetLeft(_Rectangle, -X < point.X ? point.X + X + AppModel.Current.MaskLeftWidth : AppModel.Current.MaskLeftWidth);
+                _Rectangle.Width = -X < point.X ? -X : point.X;
             }
             if (Y > 0)
             {
-                Canvas.SetTop(_Rectangle, (int)point.Y + AppModel.Current.MaskTopHeight);
-                _Rectangle.Height = Y < Height - point.Y ? Y : (int)Height - (int)point.Y;
+                Canvas.SetTop(_Rectangle, point.Y + AppModel.Current.MaskTopHeight);
+                _Rectangle.Height = Y < Height - point.Y ? Y : Height - point.Y;
             }
             else
             {
-                Canvas.SetTop(_Rectangle, -Y < point.Y ? (int)point.Y + Y + AppModel.Current.MaskTopHeight : AppModel.Current.MaskTopHeight);
-                _Rectangle.Height = -Y < point.Y ? -Y : (int)point.Y;
+                Canvas.SetTop(_Rectangle, -Y < point.Y ? point.Y + Y + AppModel.Current.MaskTopHeight : AppModel.Current.MaskTopHeight);
+                _Rectangle.Height = -Y < point.Y ? -Y : point.Y;
             }
         }
         #endregion
 
         #region 画椭圆
-        private void DrawEllipse(int X, int Y)
+        private void DrawEllipse(double X, double Y)
         {
             if (_Ellipse == null)
             {
@@ -339,29 +339,29 @@ namespace ComeCapture.Controls
             }
             if (X > 0)
             {
-                Canvas.SetLeft(_Ellipse, (int)point.X + AppModel.Current.MaskLeftWidth);
-                _Ellipse.Width = X < Width - point.X ? X : (int)Width - (int)point.X;
+                Canvas.SetLeft(_Ellipse, point.X + AppModel.Current.MaskLeftWidth);
+                _Ellipse.Width = X < Width - point.X ? X : Width - point.X;
             }
             else
             {
-                Canvas.SetLeft(_Ellipse, -X < point.X ? (int)point.X + X + AppModel.Current.MaskLeftWidth : AppModel.Current.MaskLeftWidth);
-                _Ellipse.Width = -X < point.X ? -X : (int)point.X;
+                Canvas.SetLeft(_Ellipse, -X < point.X ? point.X + X + AppModel.Current.MaskLeftWidth : AppModel.Current.MaskLeftWidth);
+                _Ellipse.Width = -X < point.X ? -X : point.X;
             }
             if (Y > 0)
             {
-                Canvas.SetTop(_Ellipse, (int)point.Y + AppModel.Current.MaskTopHeight);
-                _Ellipse.Height = Y < Height - point.Y ? Y : (int)Height - point.Y;
+                Canvas.SetTop(_Ellipse, point.Y + AppModel.Current.MaskTopHeight);
+                _Ellipse.Height = Y < Height - point.Y ? Y : Height - point.Y;
             }
             else
             {
-                Canvas.SetTop(_Ellipse, -Y < point.Y ? (int)point.Y + Y + AppModel.Current.MaskTopHeight : AppModel.Current.MaskTopHeight);
-                _Ellipse.Height = -Y < point.Y ? -Y : (int)point.Y;
+                Canvas.SetTop(_Ellipse, -Y < point.Y ? point.Y + Y + AppModel.Current.MaskTopHeight : AppModel.Current.MaskTopHeight);
+                _Ellipse.Height = -Y < point.Y ? -Y : point.Y;
             }
         }
         #endregion
 
         #region 画箭头
-        private void DrawArrow(int X, int Y)
+        private void DrawArrow(double X, double Y)
         {
             var screen = PointToScreen(point);
             if (_Arrow == null)
@@ -398,7 +398,7 @@ namespace ComeCapture.Controls
         #endregion
 
         #region 画刷
-        private void DrawLine(int X, int Y)
+        private void DrawLine(double X, double Y)
         {
             var screen = PointToScreen(point);
             if (_Line == null)
@@ -416,8 +416,8 @@ namespace ComeCapture.Controls
                 MainWindow.AddControl(_Line);
             }
             var point2 = new Point(screen.X + X, screen.Y + Y);
-            point2.X = point2.X < AppModel.Current.MaskLeftWidth ? AppModel.Current.MaskLeftWidth : point2.X > AppModel.Current.MaskLeftWidth + Width ? AppModel.Current.MaskLeftWidth + (int)Width : (int)point2.X;
-            point2.Y = point2.Y < AppModel.Current.MaskTopHeight ? AppModel.Current.MaskTopHeight : point2.Y > AppModel.Current.MaskTopHeight + Height ? AppModel.Current.MaskTopHeight + (int)Height : (int)point2.Y;
+            point2.X = point2.X < AppModel.Current.MaskLeftWidth ? AppModel.Current.MaskLeftWidth : point2.X > AppModel.Current.MaskLeftWidth + Width ? AppModel.Current.MaskLeftWidth + Width : point2.X;
+            point2.Y = point2.Y < AppModel.Current.MaskTopHeight ? AppModel.Current.MaskTopHeight : point2.Y > AppModel.Current.MaskTopHeight + Height ? AppModel.Current.MaskTopHeight + Height : point2.Y;
             points.Add(point2);
             using (var ctx = geometry.Open())
             {
@@ -468,7 +468,7 @@ namespace ComeCapture.Controls
         #endregion
 
         #region 拖动截图区域
-        private void OnMove(int X, int Y)
+        private void OnMove(double X, double Y)
         {
             #region X轴移动
             if (X > 0)
@@ -481,7 +481,7 @@ namespace ComeCapture.Controls
             }
             else
             {
-                var max = AppModel.Current.MaskLeftWidth > AppModel.Current.MaskLeftWidth + (int)Width - Limit.Right ? AppModel.Current.MaskLeftWidth + (int)Width - Limit.Right : AppModel.Current.MaskLeftWidth;
+                var max = AppModel.Current.MaskLeftWidth > AppModel.Current.MaskLeftWidth + Width - Limit.Right ? AppModel.Current.MaskLeftWidth + Width - Limit.Right : AppModel.Current.MaskLeftWidth;
                 if (-X > max)
                 {
                     X = -max;
@@ -506,7 +506,7 @@ namespace ComeCapture.Controls
             }
             else
             {
-                var max = AppModel.Current.MaskTopHeight > AppModel.Current.MaskTopHeight + (int)Height - Limit.Bottom ? AppModel.Current.MaskTopHeight + (int)Height - Limit.Bottom : AppModel.Current.MaskTopHeight;
+                var max = AppModel.Current.MaskTopHeight > AppModel.Current.MaskTopHeight + Height - Limit.Bottom ? AppModel.Current.MaskTopHeight + Height - Limit.Bottom : AppModel.Current.MaskTopHeight;
                 if (-Y > max)
                 {
                     Y = -max;
@@ -523,13 +523,13 @@ namespace ComeCapture.Controls
         #endregion
 
         #region 左缩放
-        private void Left(int X)
+        private void Left(double X)
         {
             if (X > 0)
             {
-                var max = MainWindow.Current.list.Count == 0 ? (int)Width - MainWindow.MinSize
-                    : Limit.Left - AppModel.Current.MaskLeftWidth < (int)Width - MainWindow.MinSize ? Limit.Left - AppModel.Current.MaskLeftWidth
-                    : (int)Width - MainWindow.MinSize;
+                var max = MainWindow.Current.list.Count == 0 ? Width - MainWindow.MinSize
+                    : Limit.Left - AppModel.Current.MaskLeftWidth < Width - MainWindow.MinSize ? Limit.Left - AppModel.Current.MaskLeftWidth
+                    : Width - MainWindow.MinSize;
                 if (X > max)
                 {
                     X = max;
@@ -554,7 +554,7 @@ namespace ComeCapture.Controls
         #endregion
 
         #region 右缩放
-        private void Right(int X)
+        private void Right(double X)
         {
             if (X > 0)
             {
@@ -566,9 +566,9 @@ namespace ComeCapture.Controls
             }
             else
             {
-                var max = MainWindow.Current.list.Count == 0 ? (int)Width - MainWindow.MinSize
-                    : AppModel.Current.MaskLeftWidth + (int)Width - Limit.Right < Width - MainWindow.MinSize ? AppModel.Current.MaskLeftWidth + (int)Width - Limit.Right
-                    : (int)Width - MainWindow.MinSize;
+                var max = MainWindow.Current.list.Count == 0 ? Width - MainWindow.MinSize
+                    : AppModel.Current.MaskLeftWidth + Width - Limit.Right < Width - MainWindow.MinSize ? AppModel.Current.MaskLeftWidth + Width - Limit.Right
+                    : Width - MainWindow.MinSize;
                 if (-X > max)
                 {
                     X = -max;
@@ -584,13 +584,13 @@ namespace ComeCapture.Controls
         #endregion
 
         #region 上缩放
-        private void Top(int Y)
+        private void Top(double Y)
         {
             if (Y > 0)
             {
-                var max = MainWindow.Current.list.Count == 0 ? (int)Height - MainWindow.MinSize
-                    : Limit.Top - AppModel.Current.MaskTopHeight < (int)Height - MainWindow.MinSize ? Limit.Top - AppModel.Current.MaskTopHeight
-                    : (int)Height - MainWindow.MinSize;
+                var max = MainWindow.Current.list.Count == 0 ? Height - MainWindow.MinSize
+                    : Limit.Top - AppModel.Current.MaskTopHeight < Height - MainWindow.MinSize ? Limit.Top - AppModel.Current.MaskTopHeight
+                    : Height - MainWindow.MinSize;
                 if (Y > max)
                 {
                     Y = max;
@@ -614,7 +614,7 @@ namespace ComeCapture.Controls
         #endregion
 
         #region 下缩放
-        private void Bottom(int Y)
+        private void Bottom(double Y)
         {
             if (Y > 0)
             {
@@ -626,9 +626,9 @@ namespace ComeCapture.Controls
             }
             else
             {
-                var max = MainWindow.Current.list.Count == 0 ? (int)Height - MainWindow.MinSize
-                    : AppModel.Current.MaskTopHeight + (int)Height - Limit.Bottom < (int)Height - MainWindow.MinSize ? AppModel.Current.MaskTopHeight + (int)Height - Limit.Bottom
-                    : (int)Height - MainWindow.MinSize;
+                var max = MainWindow.Current.list.Count == 0 ? Height - MainWindow.MinSize
+                    : AppModel.Current.MaskTopHeight + Height - Limit.Bottom < Height - MainWindow.MinSize ? AppModel.Current.MaskTopHeight + Height - Limit.Bottom
+                    : Height - MainWindow.MinSize;
                 if (-Y > max)
                 {
                     Y = -max;
@@ -646,12 +646,12 @@ namespace ComeCapture.Controls
         private void OnMove(object sender, MouseEventArgs e)
         {
             var point = PointToScreen(e.GetPosition(this));
-            AppModel.Current.ShowRGB= ImageHelper.GetRGB((int)point.X, (int)point.Y);
+            AppModel.Current.ShowRGB = ImageHelper.GetRGB((int)point.X, (int)point.Y);
         }
         #endregion
 
         #region 计算图片移动的极限值
-        public void ResetLimit(int left, int top, int right, int bottom)
+        public void ResetLimit(double left, double top, double right, double bottom)
         {
             ResetLeft(left);
             ResetTop(top);
@@ -659,7 +659,7 @@ namespace ComeCapture.Controls
             ResetButtom(bottom);
         }
 
-        private void ResetLeft(int value)
+        private void ResetLeft(double value)
         {
             if (value < Limit.Left)
             {
@@ -667,7 +667,7 @@ namespace ComeCapture.Controls
             }
         }
 
-        private void ResetTop(int value)
+        private void ResetTop(double value)
         {
             if (value < Limit.Top)
             {
@@ -675,7 +675,7 @@ namespace ComeCapture.Controls
             }
         }
 
-        private void ResetRight(int value)
+        private void ResetRight(double value)
         {
             if (value > Limit.Right)
             {
@@ -683,7 +683,7 @@ namespace ComeCapture.Controls
             }
         }
 
-        private void ResetButtom(int value)
+        private void ResetButtom(double value)
         {
             if (value > Limit.Bottom)
             {
